+++ test/rubocop/cop/minitest/assert_truthy_test.rb
@@ -117,4 +117,116 @@ class AssertTruthyTest < Minitest::Test
       end
     RUBY
   end
+
+  # Refute
+  def test_adds_offense_for_use_of_refute_equal_true
+    assert_offense(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute_equal(true, somestuff)
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer using `refute(somestuff)` over `refute_equal(true, somestuff)`.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute(somestuff)
+        end
+      end
+    RUBY
+  end
+
+  def test_adds_offense_for_use_of_refute_equal_true_with_message
+    assert_offense(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute_equal(true, somestuff, 'the message')
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer using `refute(somestuff, 'the message')` over `refute_equal(true, somestuff, 'the message')`.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute(somestuff, 'the message')
+        end
+      end
+    RUBY
+  end
+
+  def test_adds_offense_for_use_of_refute_equal_with_a_method_call
+    assert_offense(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute_equal(true, obj.is_something?, 'the message')
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer using `refute(obj.is_something?, 'the message')` over `refute_equal(true, obj.is_something?, 'the message')`.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          refute(obj.is_something?, 'the message')
+        end
+      end
+    RUBY
+  end
+
+  def test_adds_offense_for_use_of_refute_equal_with_a_string_variable_in_message
+    assert_offense(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          message = 'the message'
+          refute_equal(true, somestuff, message)
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer using `refute(somestuff, message)` over `refute_equal(true, somestuff, message)`.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_something
+          message = 'the message'
+          refute(somestuff, message)
+        end
+      end
+    RUBY
+  end
+
+  def test_adds_offense_for_use_of_refute_equal_with_a_constant_in_message
+    assert_offense(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        MESSAGE = 'the message'
+
+        def test_do_something
+          refute_equal(true, somestuff, MESSAGE)
+          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer using `refute(somestuff, MESSAGE)` over `refute_equal(true, somestuff, MESSAGE)`.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        MESSAGE = 'the message'
+
+        def test_do_something
+          refute(somestuff, MESSAGE)
+        end
+      end
+    RUBY
+  end
+
+  def test_does_not_offend_if_using_refute
+    assert_no_offenses(<<~RUBY, @cop)
+      class FooTest < Minitest::Test
+        def test_do_somethingra
+          refute(somestuff)
+        end
+      end
+    RUBY
+  end
 end